Refactor Covid Container; filter overview data and award amounts by DEFC V based on query param
